paint-timing/fcp-only/fcp-video-poster.html

<!DOCTYPE html>
<head>
<title>Performance Paint Timing Test: FCP due to loaded video poster</title>
<style>#log {display: none}</style>
</head>
<body>
<script src="../resources/utils.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<video id="video" width="50" height="50"></canvas>
</body>

<footer>
<script>
  promise_test(async t => {
    try {
        assert_precondition(window.PerformancePaintTiming, "Paint Timing isn't supported.");
        await new Promise(r => window.addEventListener('load', r))
        await assertNoFirstContentfulPaint(t)
        const video = document.getElementById('video')
        const src = '../resources/circles.png'
        const image = new Image()
        image.src = src
        video.setAttribute('poster', src)
        await new Promise(resolve => {
            image.onload = async () => resolve()
        })
        await assertFirstContentfulPaint(t)
    } catch (e) {
      document.getElementById('log').style.display = 'block'
      throw e
    } finally {
      document.getElementById('log').style.display = 'block'
    }
  }, 'Video should become contentful when poster is loaded')
</script>
<footer>
</html>
